Concluindo implementação da comunicação entre as apis; Implementado a sales-api; Configurando alguns tracings na aplicação para registrar informações de requisições e repostas entre os microserviços; Configurando docker-compose para subir todos os banco e aplicações;

+++ product-api/src/main/java/br/com/cursocomunicacaomicroservico/productapigradle/modules/product/service/ProductService.java
@@ -8,9 +8,11 @@ import br.com.cursocomunicacaomicroservico.productapigradle.modules.product.mode
 import br.com.cursocomunicacaomicroservico.productapigradle.modules.product.repository.ProductRepository;
 import br.com.cursocomunicacaomicroservico.productapigradle.modules.sales.client.SalesClient;
 import br.com.cursocomunicacaomicroservico.productapigradle.modules.sales.dto.SalesConfirmationDTO;
+import br.com.cursocomunicacaomicroservico.productapigradle.modules.sales.dto.SalesProductResponse;
 import br.com.cursocomunicacaomicroservico.productapigradle.modules.sales.enums.SalesStatus;
 import br.com.cursocomunicacaomicroservico.productapigradle.modules.sales.rabbitmq.SalesConfirmationSender;
 import br.com.cursocomunicacaomicroservico.productapigradle.modules.supplier.service.SupplierService;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -19,6 +21,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
+import static br.com.cursocomunicacaomicroservico.productapigradle.config.RequestUtil.getCurrentRequest;
 import static org.springframework.util.ObjectUtils.isEmpty;
 
 @Slf4j
@@ -27,6 +30,8 @@ import static org.springframework.util.ObjectUtils.isEmpty;
 public class ProductService {
 
     private static final Integer ZERO = 0;
+    private static final String TRANSACTION_ID = "transactionid";
+    private static final String SERVICE_ID = "serviceid";
 
     private final ProductRepository productRepository;
     private final SupplierService supplierService;
@@ -115,6 +120,13 @@ public class ProductService {
 
     public SuccessResponse delete(Integer id) {
         validateInformedId(id);
+        if (!productRepository.existsById(id)) {
+            throw new ValidationException("Product does not exists.");
+        }
+        var sales = getSalesByProductId(id);
+        if (!isEmpty(sales.getSalesIds())) {
+            throw new ValidationException("Product cannot be deleted, there are sales for it.");
+        }
         productRepository.deleteById(id);
         return SuccessResponse.create("the product was deleted.");
     }
@@ -155,7 +167,8 @@ public class ProductService {
             updateStock(productStockDTO);
         } catch (Exception ex) {
             log.error("Error while trying to update stock for message with error: {}", ex.getMessage(), ex);
-            var rejectedMessage = new SalesConfirmationDTO(productStockDTO.getSalesId(), SalesStatus.RJECTED);
+            var rejectedMessage = new SalesConfirmationDTO(
+                    productStockDTO.getSalesId(), SalesStatus.RJECTED, productStockDTO.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(rejectedMessage);
         }
     }
@@ -188,7 +201,8 @@ public class ProductService {
                 });
         if (!isEmpty(productsForUpdate)) {
             productRepository.saveAll(productsForUpdate);
-            var approvedMessage = new SalesConfirmationDTO(productStockDTO.getSalesId(), SalesStatus.APPROVED);
+            var approvedMessage = new SalesConfirmationDTO(
+                    productStockDTO.getSalesId(), SalesStatus.APPROVED, productStockDTO.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(approvedMessage);
         }
     }
@@ -202,35 +216,61 @@ public class ProductService {
 
     public ProductSalesResponse findProductSales(Integer id) {
         var product = findById(id);
-        try {
-            var sales = salesClient
-                    .findSalesByProductId(product.getId())
-                    .orElseThrow(() -> new ValidationException("Sales not found by this product."));
+        var sales = getSalesByProductId(product.getId());
+        return ProductSalesResponse.of(product, sales.getSalesIds());
+    }
 
-            return ProductSalesResponse.of(product, sales.getSalesIds());
+    private SalesProductResponse getSalesByProductId(Integer productId) {
+        try {
+            var currentRequest = getCurrentRequest();
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Request to GET orders by productId with data {} | [transactionId: {} | serviceId: {}]",
+                    productId, transactionid, serviceid);
+            var response = salesClient
+                    .findSalesByProductId(productId)
+                    .orElseThrow(() -> new ValidationException("Sales not found by this product."));
+            log.info("Response to GET orders by productId with data {} | [transactionId: {} | serviceId: {}]",
+                    new ObjectMapper().writeValueAsString(response), transactionid, serviceid);
+            return response;
         } catch (Exception ex) {
-            ex.printStackTrace();
             throw new ValidationException("There was an error trying to get the product's sales.");
         }
     }
 
     public SuccessResponse checkProductsStock(ProductCheckStockRequest productCheckStockRequest) {
-        if (isEmpty(productCheckStockRequest) || isEmpty(productCheckStockRequest.getProducts())) {
-            throw new ValidationException("The request data must be informed.");
+        try {
+            var currentRequest = getCurrentRequest();
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Request to POST product stock with data {} | [transactionId: {} | serviceId: {}]",
+                    new ObjectMapper().writeValueAsString(productCheckStockRequest),
+                    transactionid,
+                    serviceid);
+            if (isEmpty(productCheckStockRequest) || isEmpty(productCheckStockRequest.getProducts())) {
+                throw new ValidationException("The request data must be informed.");
+            }
+            productCheckStockRequest
+                    .getProducts()
+                    .forEach(this::validateStock);
+            var response = SuccessResponse.create("The stock is ok.");
+            log.info("Response to POST product stock with data {} | [transactionId: {} | serviceId: {}]",
+                    new ObjectMapper().writeValueAsString(response),
+                    transactionid,
+                    serviceid);
+            return response;
+        } catch (Exception ex) {
+            throw new ValidationException(ex.getMessage());
         }
-        productCheckStockRequest
-                .getProducts()
-                .forEach(this::validateStock);
-        return SuccessResponse.create("The stock is ok.");
     }
 
     private void validateStock(ProductQuantityDTO productQuantityDTO) {
-        if (isEmpty(productQuantityDTO.getProductId()) || isEmpty(productQuantityDTO.getQuantity()))  {
+        if (isEmpty(productQuantityDTO.getProductId()) || isEmpty(productQuantityDTO.getQuantity())) {
             throw new ValidationException("Product id and quantity must be informed.");
         }
         var product = findById(productQuantityDTO.getProductId());
         if (productQuantityDTO.getQuantity() > product.getQuantityAvailable()) {
-            throw new ValidationException( String.format("The product %s is out of stock.", product.getId()));
+            throw new ValidationException(String.format("The product %s is out of stock.", product.getId()));
         }
     }
